refactor couple files to arrow module
