+++ crates/proof-of-sql/src/sql/proof/query_result.rs
@@ -3,8 +3,7 @@ use crate::base::{
     proof::ProofError,
     scalar::Scalar,
 };
-#[cfg(feature = "arrow")]
-use arrow::{error::ArrowError, record_batch::RecordBatch};
+use arrow::record_batch::RecordBatch;
 use snafu::Snafu;
 
 /// Verifiable query errors
@@ -59,15 +58,6 @@ impl<S: Scalar> QueryData<S> {
     #[must_use]
     pub fn into_record_batch(self) -> RecordBatch {
         self.try_into().unwrap()
-    }
-}
-
-#[cfg(feature = "arrow")]
-impl<S: Scalar> TryFrom<QueryData<S>> for RecordBatch {
-    type Error = ArrowError;
-
-    fn try_from(value: QueryData<S>) -> Result<Self, Self::Error> {
-        Self::try_from(value.table)
     }
 }
